Fixed some animations which were based on frame counting
They are now based on WorldTime.
At some places, the frame counter was just used as a randomizer.
Removed the MoveSceneFrame variable in this process.
There are still a lot of places which need to be adjusted, this was just the tip of the iceberg.
E.g. there are a lot of code lines like
```
float fActionSpeed = pModel->Actions[pObject->CurrentAction].PlaySpeed;
float fSpeedPerFrame = fActionSpeed / 10.f;
float fAnimationFrame = pObject->AnimationFrame - fActionSpeed;
```

or places like in CSParts where the animation speed is defined as Velocity, which gets applied on each frame.