current state of unicode conversion
This will be a big step forward, when it's done.
Unfortunately, it's not finished yet. 8 little compiler errors are left :)
After fixing that, it needs some testing to find potential new issues... I'm not sure if these changes are all correct, especially when reading bmd files with strings.

Sending the packets need to be adapted as well to the new send functions, to make it work properly, of course.